requirements.txt
fastapi>=0.110.0
pydantic>=2.0.0
uvicorn[standard]>=0.23.0
streamlit>=1.29.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
requests>=2.31.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0
shap>=0.43.0
lime>=0.2.0.1
xgboost>=1.7.0
hopsworks[python]>=4.2,<4.3; python_version < '3.12'
python-dotenv>=1.0.0

